// api/tmrev-api-v2/schemas/feedItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * TMREV API v2
 * Review aggregation and movie database API
 * OpenAPI spec version: 2.0.0
 */
import type { AdvancedScore } from './advancedScore';
import type { FeedItemVotes } from './feedItemVotes';
import type { FeedItemMovieDetails } from './feedItemMovieDetails';
import type { FeedItemProfile } from './feedItemProfile';

export interface FeedItem {
	_id: string;
	userId: string;
	tmdbID: number;
	title: string;
	notes: string;
	public: boolean;
	averagedAdvancedScore: number;
	advancedScore: AdvancedScore;
	votes: FeedItemVotes;
	createdAt: string;
	updatedAt: string;
	reviewedDate: string;
	movieDetails: FeedItemMovieDetails;
	commentCount: number;
	profile: FeedItemProfile;
	/** Whether the authenticated user has upvoted this review */
	hasUpvoted: boolean;
	/** Whether the authenticated user has downvoted this review */
	hasDownvoted: boolean;
}